added dockerfile, wsgi file structure, work in progress
