+++ .gitignore
@@ -1,3 +1,13 @@
-Dockerfile
-.wsgi.py
-.Procfile
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# Environments
+.env
+.venv
+env/
+venv/
+ENV/
+env.bak/
+venv.bak/
