+++ README.md
@@ -1,2 +1,4 @@
 # flask-CRUD-app
 An app to keep track of items while I move house.
+Served by gunicorn
+Containerized
