+++ wsgi.py
@@ -1,4 +1,4 @@
-from app import server
+# from app import server
 
-if __name__ == "__main__":
-    server.run(host='0.0.0.0', port=8000)
+# if __name__ == "__main__":
+#     server.run(host='0.0.0.0', port=8000)
